Give client events the refactor
Move all client event imports out of __init__ and go for more direct imports
