--- discordbot/clienteventclasses/__init__.py
@@ -1,19 +1,0 @@
-
-"""
-clienteventclasses/ contains all of the classes for different client events
-Each file contains a different class for handling an event that isn't dissimilar
-to the name of the file
-"""
-
-from .ondirectmessage import OnDirectMessage
-from .onemojicreate import OnEmojiCreate
-from .onmemberjoin import OnMemberJoin
-from .onmemberleave import OnMemberLeave
-from .onmessage import OnMessage
-from .onmessageedit import OnMessageEdit
-from .onreactionadd import OnReactionAdd
-from .onready import OnReadyEvent
-from .onstickercreate import OnStickerCreate
-from .onthreadcreate import OnThreadCreate
-from .onthreadupdate import OnThreadUpdate
-from .onvoicestatechange import OnVoiceStateChange
